# Conflicts:
#	Solutions/win10-dll/win10-dll.vcxproj
#	Solutions/win32-dll/win32-dll.vcxproj

+++ lib/http/HttpClient_WinRt.cpp
@@ -97,7 +97,7 @@ namespace ARIASDK_NS_BEGIN {
 								
 			// Initialize the in-memory stream where data will be stored.
 			DataWriter^ dataWriter = ref new DataWriter();
-			dataWriter->WriteBytes((Platform::ArrayReference<unsigned char>(reinterpret_cast<unsigned char*>(request->m_body.data()), request->m_body.size())));
+			dataWriter->WriteBytes((Platform::ArrayReference<unsigned char>(reinterpret_cast<unsigned char*>(request->m_body.data()), (DWORD)request->m_body.size())));
 			IBuffer ^ibuffer = dataWriter->DetachBuffer();
 			HttpBufferContent^ httpBufferContent = ref new HttpBufferContent(ibuffer);
 
@@ -196,7 +196,7 @@ namespace ARIASDK_NS_BEGIN {
 					{
 						response->m_body.reserve(length);
 						unsigned char* storage = static_cast<unsigned char*>(response->m_body.data());
-						DataReader::FromBuffer(buffer)->ReadBytes(ArrayReference<unsigned char>(storage, length));
+						DataReader::FromBuffer(buffer)->ReadBytes(ArrayReference<unsigned char>(storage, (DWORD)length));
 						//response->m_body = std::string(reinterpret_cast<char*>(storage), length);
 					}
 				}
